requirements.txt
# MangroVision - Complete Dependencies
# Python 3.12+ compatible

# Core UI Framework
streamlit>=1.30.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# GIS and Geometry Processing
shapely>=2.0.0
geopandas>=0.14.0
pyproj>=3.6.0

# Data Processing
pandas>=2.0.0

# Backend API (for map-based detection)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Utilities (already installed but included for completeness)
tqdm>=4.66.0
typer>=0.9.0
click>=8.1.0
rich>=13.0.0

# Optional but recommended for better performance
matplotlib>=3.7.0  # For visualization if needed
rasterio>=1.3.0    # For orthophoto metadata extraction (optional)
gdal>=3.7.0        # For GIS operations (optional, can be tricky to install)
